Changements de nom de variable, implémentation de l'interface Graphe, et corrections des methodes de GrapheListe

+++ src/main/java/Arc.java
@@ -1,13 +1,15 @@
 public class Arc {
     private String destination;
     private double cout;
-    public Arc(String d, int c){
+
+    public Arc(String d, double c){
         this.destination = d;
         this.cout = c;
     }
     public String getDestination(){
         return this.destination;
     }
+
     public double getCout(){
         return this.cout;
     }
